+++ primitives/src/big_num.rs
@@ -222,6 +222,55 @@ where
     serializer.serialize_str(&num.to_str_radix(10))
 }
 
+#[cfg(feature = "postgres")]
+pub mod postgres {
+    use super::BigNum;
+    use bytes::BytesMut;
+    use postgres_types::{FromSql, IsNull, ToSql, Type};
+    use std::error::Error;
+
+    impl<'a> FromSql<'a> for BigNum {
+        fn from_sql(ty: &Type, raw: &'a [u8]) -> Result<BigNum, Box<dyn Error + Sync + Send>> {
+            use std::convert::TryInto;
+
+            let str_slice = <&str as FromSql>::from_sql(ty, raw)?;
+
+            Ok(str_slice.try_into()?)
+        }
+
+        fn accepts(ty: &Type) -> bool {
+            match *ty {
+                Type::TEXT | Type::VARCHAR => true,
+                _ => false,
+            }
+        }
+    }
+
+    impl ToSql for BigNum {
+        fn to_sql(
+            &self,
+            ty: &Type,
+            w: &mut BytesMut,
+        ) -> Result<IsNull, Box<dyn Error + Sync + Send>> {
+            <String as ToSql>::to_sql(&self.0.to_string(), ty, w)
+        }
+
+        fn accepts(ty: &Type) -> bool {
+            match *ty {
+                Type::TEXT | Type::VARCHAR => true,
+                _ => false,
+            }
+        }
+
+        fn to_sql_checked(
+            &self,
+            ty: &Type,
+            out: &mut BytesMut,
+        ) -> Result<IsNull, Box<dyn Error + Sync + Send>> {
+            <String as ToSql>::to_sql_checked(&self.0.to_string(), ty, out)
+        }
+    }
+}
 #[cfg(test)]
 mod test {
     use super::*;
